refactor(nova-guard): one admission wire contract, not two copies
HANDOFF item 4. classify_admit, the request and settlement bodies and the
URL shaping existed twice: once in policy/admission.rs (native, reqwest) and
once in policy/worker_remote.rs (wasm, worker::Fetch). They were duplicated
because both native modules are reqwest-bound at the MODULE level, and the
justification recorded at the time was that tests on both sides assert the
same contract points "so the two copies cannot silently diverge".

That reasoning does not hold. Paired tests catch a divergence someone wrote
a test for; they cannot catch the failure that actually matters here, which
is one copy being edited and the other not. Nothing fails, both suites stay
green, and the edge quietly bills differently from the native server. The
new pricingVersion field landing in the next commit is exactly that shape of
change: one field, two places, no compiler and no test to notice if only one
gets it.

src/policy/admission_wire.rs is now the single definition of the wire
contract - BlockedDecision (and its rendering as a gateway PolicyDecision,
so a platform block looks byte-identical at the edge and natively),
Reservation, Admission, AdmitRequest, SettlementUsage, Settlement,
classify_admit, sanitize_cost, urlencode and truncate_body. It pulls in no
reqwest, tokio or chrono, so it compiles for wasm32; that is also why
Settlement::to_json takes its timestamp as an argument rather than reading a
clock, with the native side passing chrono and the Worker passing Date.now().

Both clients re-export from it. remote.rs's truncate_body was a third copy
of the same function and now re-exports too. 791 lines deleted, 25 added.

What deliberately did NOT move: WorkerRemoteConfig::from_values. It mirrors
the native RemoteConfig::from_values because the two must AGREE on what a
half-applied secret set means, not because they are the same code - the
Worker reads env.secret()/env.var() and refuses shared tenancy, the native
side parses process env and implements it. Sharing them would couple two
things that are only required to reach the same verdict.

The duplicated contract tests were merged into the new module rather than
kept on both sides. Native-only tests stayed where they are: URL shaping,
strict-mode selection, enforcement env parsing, and the settlement Drop
behaviour.

This is a pure move: no behaviour change is intended, and the live workerd
E2E from the previous commit re-run unchanged is the evidence, since it
exercises the moved code through the real transport rather than through the
tests that moved with it.

Verified:
  cargo fmt --all -- --check                                    clean
  cargo clippy --all-targets --all-features -- -D warnings      clean
  cargo clippy --target wasm32-unknown-unknown --lib ... -D warnings  clean
  worker-build --release                                        ok
  scripts/novaguard_worker_e2e.sh            17/17, all phases passed
  test result: ok. 396 passed; 0 failed; ...   (lib, was 395)
  test result: ok. 51 passed; 0 failed; ...    (novaguard_platform)
  test result: ok. 12 passed; 0 failed; ...    (policy_integration)
